Delete Kkanburis_ver3.py and Modify Kkanburis_ver2.py
Kkanburis_ver3.py 파일 내용  Kkanburis_ver2.py 으로 옮겼습니다!

[변경할 내용 추가]

1. 동일한 사용자 login 시 중복 처리

-> 점수 입력 및 데이터베이스에 등록 시 동일한 사용자의 경우 더 높은 점수를 기준으로 업데이트 하기
