+++ Kkanburis/Kkanburis_ver2.py
@@ -14,9 +14,8 @@ pygame.init()
 
 # Define values
 color_active = pygame.Color('lightskyblue3')
-color_inactive = pygame.Color('blue')
+color_inactive = pygame.Color('dodgerblue2')
 color = color_inactive
-# text=''
 FONT = pygame.font.Font(None, 32)
 
 block_size = 17  # Height, width of single block
@@ -168,15 +167,22 @@ class button():   # 버튼 객체
                 return True
         return False
 
+    def isOver_2(self, pos):
+        #start 화면에서 single,pvp,help,setting을 위해서 y좌표 좁게 인식하도록
+        if pos[0] > self.x - (self.width / 2) and pos[0] < self.x + (self.width / 2):
+            if pos[1] > self.y - (self.height / 4) and pos[1] < self.y + (self.height / 4):#243줄에서의 2을 4로 바꿔주면서 좁게 인식할수 있도록함. 더 좁게 인식하고 싶으면 숫자 늘려주기#
+                return True
+        return False
+
 # InoutBox 초기 설정
-FONT = pygame.font.Font(None, 32)
+# FONT = pygame.font.Font(None, 32)
+# text = ''
 class InputBox:
     def __init__(self, x, y, w, h, text = ''):
         self.rect = pygame.Rect(x, y, w, h)
         self.color = color_inactive
         self.text = text
         self.txt_surface = FONT.render(text, True, self.color)   # FONT = pygame.font.Font(None, 32)
-        # self.txt_surface = pygame.font.Font.render(text, True, self.color)
         self.active = False
 
     def handle_event(self, event):
@@ -199,28 +205,27 @@ class InputBox:
                     self.text = self.text[:-1]
                 else:
                     self.text += event.unicode
-                # Re-render the text
-                # text_item = ui_variables.h5.render("ITEM", 1, ui_variables.real_white)
-                self.text_surface = FONT.render(text, True, self.color)
+                # Re-render the text.
+                self.txt_surface = FONT.render(self.text, True, self.color)
 
     def update(self):
-        # Resize the box if the txt is too long
-        width = max(200, self.txt_surface.get_width() + 10)
+        # Resize the box if the text is too long.
+        width = max(200, self.txt_surface.get_width()+10)
         self.rect.w = width
 
     def draw(self, screen):
         # Blit the text
         screen.blit(self.txt_surface, (self.rect.x+5, self.rect.y+5))
         # Blit the text
-        screen.draw.rect(screen, self.color, self.rect, 2)
+        pygame.draw.rect(screen, self.color, self.rect, 2)
 
 # InputBox 설정
 # Sign up에서 box
-input_box1 = InputBox(int(board_width * 322 / 800), int(board_height * 155.5/450), 156, 32)
+input_box1 = InputBox(int(board_width * 322 / 800), int(board_height * 195.5/450), 156, 32)
 input_box2 = InputBox(int(board_width * 322 / 800), int(board_height * 242.5/450), 156, 32)
 input_boxes_signup = [input_box1, input_box2]
 # Sign in에서 box
-input_box3 = InputBox(int(board_width * 322 / 800), int(board_height * 155.5/450), 156, 32)
+input_box3 = InputBox(int(board_width * 322 / 800), int(board_height * 195.5/450), 156, 32)
 input_box4 = InputBox(int(board_width * 322 / 800), int(board_height * 242.5/450), 156, 32)
 input_boxes_signin = [input_box3, input_box4]
 
@@ -228,7 +233,7 @@ input_boxes_signin = [input_box3, input_box4]
 # help_image = 'assets/images/help.png'
 # start_button = button(board_width * 0.5, board_height * 0.5, 146, 43, 1, start_image)
 
-background_image = 'assets/vector/Background.png'   # 배경 보드 이미지 
+background_image = 'assets/vector/Background.png'   # 배경 보드 이미지  
 
 single_button_image = 'assets/vector/single_button.png'   # 싱글 모드 버튼 이미지
 clicked_single_button_image = 'assets/vector/clicked_single_button.png'   # 싱글 모드 버튼 클릭 시 이미지
@@ -311,12 +316,13 @@ log_board = 'assets/vector/log_or_sign_board.png'
 button_log_back = 'assets/vector/button_l_back.png'   # 이거 추가
 button_log_back_clicked = 'assets/vector/button_l_back_clicked.png'   # 이거 추가
 button_sign_up = 'assets/vector/button_sign_up.png'
-# button_sign_up = 'Kkanburis/assets/vector/button_sign_up.png'
 button_sign_up_clicked = 'assets/vector/button_sign_up_clicked.png'
 button_sign_in = 'assets/vector/button_sign_in.png'
 button_sign_in_clicked = 'assets/vector/button_sign_in_clicked.png'
-button_quit = 'assets/vector/button_quit.png'
-button_quit_clicked = 'assets/vector/button_quit_clicked.png'
+# button_quit = 'assets/vector/button_quit.png'
+button_quit = 'assets/vector/quit_button.png'
+# button_quit_clicked = 'assets/vector/button_quit_clicked.png'
+button_quit_clicked = 'assets/vector/clicked_quit_button.png'
 
 # 버튼 객체 생성 - class button()에서 확인
 # def __init__(self, board_width, board_height, x_rate, y_rate, width_rate, height_rate, img = '')
@@ -365,15 +371,15 @@ midiumsize_check_button = button(board_width, board_height, 0.5, 0.45, 0.18, 0.1
 bigsize_check_button = button(board_width, board_height, 0.5, 0.65, 0.18, 0.14, 1, bigsize_board)
 
 # 회원가입/로그인 버튼
-sign_up_button1 = button(board_width, board_height, 0.415, 0.5, 12/80, 4/45, button_sign_up)
-sign_in_button1 = button(board_width, board_height, 0.585, 0.5, 12/80, 4/45, button_sign_in)
-log_quit = button(board_width, board_height, 0.5, 0.9, 0.16, 0.084, button_quit)
+sign_up_button1 = button(board_width, board_height, 0.415, 0.5, 12/80, 4/45, 1, button_sign_up)
+sign_in_button1 = button(board_width, board_height, 0.585, 0.5, 12/80, 4/45, 1, button_sign_in)
+log_quit = button(board_width, board_height, 0.5, 0.9, 0.16, 0.084, 1, button_quit)
 # log_quit = button(board_width, board_height, 0.5, 0.9, 0.16, 0.084, quit_button_image)
 
 # login page 2) sign up / sign in
-sign_up_button2 = button(board_width, board_height, 0.415, 0.7, 12/80, 4/45, button_sign_up)
-sign_in_button2 = button(board_width, board_height, 0.415, 0.7, 12/80, 4/45, button_sign_in)
-log_back = button(board_width, board_height, 0.585, 0.7, 12/80, 4/45, button_log_back)
+sign_up_button2 = button(board_width, board_height, 0.415, 0.7, 12/80, 4/45, 1, button_sign_up)
+sign_in_button2 = button(board_width, board_height, 0.415, 0.7, 12/80, 4/45, 1, button_sign_in)
+log_back = button(board_width, board_height, 0.585, 0.7, 12/80, 4/45, 1, button_log_back)
 
 # 아이템을 버튼 클래스 객체로 생성 - 비율은 테트리스 블럭 하나 사이즈 -> block_size = int(board_height * 0.045) 만큼으로
 # bomb = button()
@@ -732,7 +738,6 @@ def erase_mino(x, y, mino, r):
                 matrix[x + j][y + i] = 0   # 해당 위치에 블록을 없애 빈 곳으로 만들기
                 
 
-
 def erase_mino_2P(x, y, mino, r):
     grid = tetrimino.mino_map[mino - 1][r]
 
@@ -943,22 +948,6 @@ def set_vol(val):
     print(volume)
     ui_variables.click_sound.set_volume(volume)
 
-# # Erase a tetrimino
-# def erase_mino(x, y, mino, r):
-#     grid = tetrimino.mino_map[mino - 1][r]
-
-#     # Erase ghost
-#     for j in range(21):
-#         for i in range(10):
-#             if matrix[i][j] == 8:   # 테트리스 블록에서 해당 행렬 위치에 ghost 블록이 존재하면,
-#                 matrix[i][j] = 0   # 없애서 빈 곳으로 만들기
- 
-#     # Erase mino
-#     for i in range(mino_matrix_y):
-#         for j in range(mino_matrix_x):
-#             if grid[i][j] != 0:   # 테트리스 블록에서 해당 행렬 위치에 블록이 존재하면, 
-#                 matrix[x + j][y + i] = 0   # 해당 위치에 블록을 없애 빈 곳으로 만들기
-
 # 아이템 획득 - 콤보 11 달성 시 item_list 중 랜덤으로 
 def get_item():    # inventory_list에 아이템 생성
     if len(inventory_list) < 3:
@@ -987,19 +976,12 @@ def use_item():
 def use_bomb(x, y, mino, r):   # 행 삭제 폭탄
     grid = tetrimino.mino_map[mino - 1][r]
 
-    k = height   # 마지막 행 제거
-
-    while k > 0:
-        for i in range(width):
-            matrix[i][k] = matrix[i][k-1]   # 지워진 블록 윗 줄을 한 줄 아래로 내리기
-        k -= 1
-
-    # for i in range(3):
-    #     k = height 
-    #     while k > 0:
-    #         for i in range(width):
-    #             matrix[i][k] = matrix[i][k-1]
-    #         k -= 1
+    for i in range(5):   # bomb 아이템 사용 시 5줄 제거
+        k = height 
+        while k > 0:
+            for i in range(width):
+                matrix[i][k] = matrix[i][k-1]   # 지워진 블록 윗 줄을 한 줄 아래로 내리기
+            k -= 1
 
 # def use_bomb(x, y, mino, r):   # 3*3, 5*5 삭제 폭탄
 #     grid = tetrimino.mino_map[mino - 1][r]
@@ -1054,7 +1036,6 @@ main = False
 combo_count = 0
 combo_count_2P = 0   # pvp 모드에서 2P의 콤보 처리를 위해 추가 
 score = 0
-# score2 = 0   # -> score_2P로 통일
 score_2P = 0
 level = 1
 level_2P = 1
@@ -1106,15 +1087,11 @@ bomb_size = 3
 item_size = 30
 
 # inventory 출력을 위한 위치 변수
-sidebar_width = int(board_width * 0.5312)
-# dx = int(board_width * 0.045) + sidebar_width + block_size * j 
+sidebar_width = int(board_width * 0.5312) 
 dx_inventory1 = int(board_width * 0.135) + sidebar_width
 dx_inventory2 = int(board_width * 0.175) + sidebar_width
 dx_inventory3 = int(board_width * 0.215) + sidebar_width
 dx_inventory = [dx_inventory1, dx_inventory2, dx_inventory3]
- # score value 위치: (int(board_width * 0.055) + sidebar_width, int(board_height * 0.5614)
- # Level value 위치: int(board_widt * 0.7219)
- # text item 위치: int(board_height * 0.6791)
 dy_inventory = int(board_height * 0.75)
 
 # 아이템 이미지 scale
@@ -1268,6 +1245,23 @@ while not done:
                 set_volume()
 
             # elif event.type == VIDEORESIZE: 부분 추가
+            elif event.type == VIDEORESIZE:
+                board_width = event.w
+                board_height = event.h
+                if board_width < min_width or board_height < min_height: #최소 너비 또는 높이를 설정하려는 경우
+                    board_width = min_width
+                    board_height = min_height
+                if not ((board_rate-0.1) < (board_height/board_width) < (board_rate+0.05)): #높이 또는 너비가 비율의 일정수준 이상을 넘어서게 되면
+                    board_width = int(board_height / board_rate) #너비를 적정 비율로 바꿔줌
+                    board_height = int(board_width*board_rate) #높이를 적정 비율로 바꿔줌
+
+
+                block_size = int(board_height * 0.045) #블록 크기비율 고정
+                screen = pygame.display.set_mode((board_width, board_height), pygame.RESIZABLE)
+
+                for i in range(len(button_list)):
+                        button_list[i].change(board_width, board_height)
+
 
     # 화면(크기) 설정 창  ->  각 화면 크기 설정 시 화면(board) 크기 설정 후 -> board 사이즈에 비례하게 크기 각각 조절 -> 조절하는 비율 변수화하기 ?
     elif screen_setting:
@@ -2773,6 +2767,7 @@ while not done:
         pygame.display.update()
 
     elif signup:
+        screen.fill(ui_variables.white)
         draw_image(screen, login_bg, board_width * 0.5, board_height * 0.5, board_width, board_height)
         draw_image(screen, signup_board, board_width * 0.5, board_height * 0.55, int(board_width * 3/8), int(board_height * 24/45))
         
@@ -2793,23 +2788,23 @@ while not done:
             if event.type == USEREVENT:
                 pygame.time.set_timer(pygame.USEREVENT, set_300)
             elif event.type == pygame.MOUSEMOTION:
-                if sign_up_button2.isOver_2(pos):
+                if sign_up_button2.isOver(pos):
                     sign_up_button2.image = button_sign_up_clicked
                 else:
                     sign_up_button2.image = button_sign_up
-                if log_back.isOver_2(pos):
+                if log_back.isOver(pos):
                     log_back.image = button_log_back_clicked
                 else:
                     log_back.image = button_log_back
             elif event.type == pygame.MOUSEBUTTONDOWN:
-                if sign_up_button2.isOver_2(pos):
+                if sign_up_button2.isOver(pos):
                     ui_variables.click_sound.play()
                     id_text = input_box1.text
                     pw_text = input_box2.text
                     add_id(id_text)
                     add_pw(id_text, pw_text)
                     signup = False
-                if log_back.isOver_2(pos):
+                if log_back.isOver(pos):
                     ui_variables.click_sound.play()
                     signup = False
             elif event.type == VIDEORESIZE:
@@ -2829,6 +2824,7 @@ while not done:
                         button_list[i].change(board_width, board_height)
 
     elif signin:
+        screen.fill(ui_variables.white)
         draw_image(screen, login_bg, board_width*0.5, board_height*0.5,
         board_width, board_height)
         draw_image(screen, signin_board, board_width*0.5, board_height*0.55,
@@ -2846,34 +2842,36 @@ while not done:
             for box in input_boxes_signin:
                 box.draw(screen)
             pygame.display.update()
+
             if event.type == USEREVENT:
                 pygame.time.set_timer(pygame.USEREVENT, set_300)
                 pygame.display.update()
             elif event.type == pygame.MOUSEMOTION:
-                if sign_in_button2.isOver_2(pos):
+                if sign_in_button2.isOver(pos):
                     sign_in_button2.image = button_sign_in_clicked
                 else:
                     sign_in_button2.image = button_sign_in
-                if log_back.isOver_2(pos):
+                if log_back.isOver(pos):
                     log_back.image = button_log_back_clicked
                 else:
                     log_back.image = button_log_back
 
             elif event.type == pygame.MOUSEBUTTONDOWN:
-                if sign_in_button2.isOver_2(pos):
+                if sign_in_button2.isOver(pos):
                     ui_variables.click_sound.play()
                     id_text = input_box3.text
                     pw_text = input_box4.text
-                    if check_info(id_text, pw_text):
-                        signin= False
-                        main = True
-                        # num_earthquake = load_earthquake_data(id_text)
-                        # num_light = load_light_data(id_text)
-                        # num_tnt = load_tnt_data(id_text)
-                        # gold = load_gold_data(id_text)
-                        user_id = id_info(id_text)
 
-                if log_back.isOver_2(pos):
+                    signin = False
+                    main = True
+                    user_id = id_info(id_text)   # 확인
+
+                    # if check_info(id_text, pw_text):   # 우선, 버튼 넘어가는 것을 확인하기 위해 주석 처리 -> database.py에서 수정
+                    #     signin= False
+                    #     main = True
+                    #     user_id = id_info(id_text)
+
+                if log_back.isOver(pos):
                     ui_variables.click_sound.play()
                     signin = False
 
@@ -2919,9 +2917,11 @@ while not done:
 
             elif event.type == KEYDOWN:                
                   
-
                 if event.key == K_RETURN:
                     ui_variables.click_sound.play()
+
+                    add_score(game_status, user_id, score)
+
                     '''
                     outfile = open('leaderboard.txt', 'a')
                     outfile.write( text + ' ' + str(score) + '\n')
@@ -2998,6 +2998,9 @@ while not done:
                 if ok_button.isOver(pos):
                     ui_variables.click_sound.play()
                     ui_variables.click_sound.play()
+
+                    add_score(game_status, user_id, score)
+
                     '''
                     outfile = open('leaderboard.txt', 'a')
                     outfile.write(text+ ' ' + str(score) + '\n')
@@ -3193,6 +3196,12 @@ while not done:
             elif event.type == KEYDOWN:            
                 if event.key == K_RETURN:
                     ui_variables.click_sound.play()
+
+                    if score > score_2P:
+                        add_score(game_status, user_id, score)
+                    else:
+                        add_score(game_status, user_id, score_2P)
+
                     '''
                     if score > score_2P:
                         outfile = open('leaderboard.txt', 'a')
@@ -3272,13 +3281,18 @@ while not done:
                     ui_variables.click_sound.play()
 
                     if score > score_2P:
-                        outfile = open('leaderboard.txt', 'a')
-                        outfile.write( text + ' ' + str(score) + '\n')
-                        outfile.close()
+                        add_score(game_status, user_id, score)
                     else:
-                        outfile = open('leaderboard.txt', 'a')
-                        outfile.write( text + ' ' + str(score_2P) + '\n')
-                        outfile.close()
+                        add_score(game_status, user_id, score_2P)
+
+                    # if score > score_2P:
+                    #     outfile = open('leaderboard.txt', 'a')
+                    #     outfile.write( text + ' ' + str(score) + '\n')
+                    #     outfile.close()
+                    # else:
+                    #     outfile = open('leaderboard.txt', 'a')
+                    #     outfile.write( text + ' ' + str(score_2P) + '\n')
+                    #     outfile.close()
                     
                     text=''
                     game_over = False
@@ -3434,7 +3448,11 @@ while not done:
 
     # Start screen  ->  여기가 기존에서는 메인
     elif main:
-        main = True
+        signin = False
+        signup = False
+        text=''
+
+        # 변수 초기화 부분 추가
 
         for event in pygame.event.get():
             pos = pygame.mouse.get_pos()
@@ -3539,16 +3557,16 @@ while not done:
 
         if not start:
             pygame.display.update()
-            clock.tick(3)
 
     else:
+        screen.fill(ui_variables.white)
         draw_image(screen, login_bg, board_width*0.5, board_height*0.5,
         board_width, board_height)
         draw_image(screen, log_board, board_width*0.5, board_height*0.5,
         int(board_width*3/8), int(board_height*2/9))
-        # sign_up_button1.draw(screen,(0,0,0))
-        # sign_in_button1.draw(screen,(0,0,0))
-        # log_quit.draw(screen,(0,0,0))
+        sign_up_button1.draw(screen,(0,0,0))
+        sign_in_button1.draw(screen,(0,0,0))
+        log_quit.draw(screen,(0,0,0))
 
         for event in pygame.event.get():
             pos = pygame.mouse.get_pos()
@@ -3558,27 +3576,27 @@ while not done:
                 pygame.time.set_timer(pygame.USEREVENT, 300)
                 pygame.display.update()
             elif event.type == pygame.MOUSEMOTION:
-                if sign_up_button1.isOver_2(pos):
+                if sign_up_button1.isOver(pos):
                     sign_up_button1.image = button_sign_up_clicked
                 else:
                     sign_up_button1.image = button_sign_up
-                if sign_in_button1.isOver_2(pos):
+                if sign_in_button1.isOver(pos):
                     sign_in_button1.image = button_sign_in_clicked
                 else:
                     sign_in_button1.image = button_sign_in
-                if log_quit.isOver_2(pos):
+                if log_quit.isOver(pos):
                     log_quit.image = button_quit_clicked
                 else:
                     log_quit.image = button_quit
 
             elif event.type == pygame.MOUSEBUTTONDOWN:
-                if sign_up_button1.isOver_2(pos):
+                if sign_up_button1.isOver(pos):
                     ui_variables.click_sound.play()
                     signup = True
-                if sign_in_button1.isOver_2(pos):
+                if sign_in_button1.isOver(pos):
                     ui_variables.click_sound.play()
                     signin = True
-                if log_quit.isOver_2(pos):
+                if log_quit.isOver(pos):
                     ui_variables.click_sound.play()
                     done = True
             elif event.type == VIDEORESIZE:
